Removed the weird cast to short

+++ src/Core/Src/main.c
@@ -318,7 +318,7 @@ int cast(void) {
   for (int rCount = 0; rCount < RAYS; rCount++) {
     // This uses David Ziemkiewicz' method of velocities and times, as well as Lodev's DDA. Massive thanks to both of these legends.
     float       rVelocityX = _cosLUT[rAngle], rVelocityY = _sinLUT[rAngle], rIntersectX = _pPosX, rIntersectY = _pPosY, rTimeX, rTimeY, rLength = 0;
-    int         mX = (short)rIntersectX, mY = (short)rIntersectY; // Casting to short instead of int is faster for some unknown reason.
+    int         mX = (int)rIntersectX, mY = (int)rIntersectY;
     _Bool       rVelSignX = (rVelocityX > 0), rVelSignY = (rVelocityY > 0);
     signed char rStepX = rVelSignX ? 1 : -1, rStepY = rVelSignY ? 1 : -1, rHitSide = 0;
